Move fallback logic into own file
